requirements.txt
streamlit
supabase
python-dotenv
yt-dlp